Refine FixedSizeListBuilder
